Add Convertible protocol support to Function arguments
For argument providers which cannot (necessarily) supply values which
are directly castable from Any to T without knowing in advance what T
is, and thus supply a proxy value conforming to Convertible instead.

The reason for this is that in Lua, a string is used to represent both
String and Data, and tables are used for both Array and Dictionary,
thus the correct object type cannot be constructed until the Swift type
is known, inside the perform block.