fix: Build the command-line interface using an Xcode project
Embedding an Info.plist, required for notarization, is not really supported with the Swift Package Manager, necessitating a whole bunch of custom linker flags and custom build commands to pre-process the file to include version numbers. Since it's abundantly clear that this isn't a use-case Apple support right now, this change moves the cli into an Xcode project and leaves fat binary generation, Info.plist embedding and signing to Xcode.

This is a deeply disappointing outcome as I have hopes to support Linux in the future but while macOS is the only supported platform, this feels like a necessary short-term compromise.