fix: Clean up the import configuration
This change adopts the type-safe configuration for most of 'site.yaml' and cleans up the definition of the importer steps. It represents a breaking change to the configuration and, as such, bumps the configuration version to 2. This is preparatory work for extracting path generation and defaults to a top-level structure that can be shared between importers to simplify settings.