ci: Work-around build issues relating to duplicate CLua definitions (#298)
It looks like there's a bug in the latest SwiftPM which causes it to
generate two module maps for CLua (one for the tool, and one for the
library) and then, in the outer package, see those as duplicate
definitions. Xcode still manages to build the project just fine though
(typical), so this change disables the `swift build` and `swift test`
steps, leaving the work to Xcode which we're already using to build and
package the InContext helper and CLI. Of course it sucks because it
means we loose all our automated testing, but it's better than a project
that doesn't build at all.